refactor to avoid circular import
