+++ src/cloudWatchLogs/document/logStreamDocumentProvider.ts
@@ -11,11 +11,7 @@ import {
     getInitialLogData,
 } from '../registry/logStreamRegistry'
 import { getLogger } from '../../shared/logger'
-import { uriToKey, findOccurencesOf, parseCloudWatchLogsUri, createURIFromArgs } from '../cloudWatchLogsUtils'
-
-const HIGHLIGHTER = vscode.window.createTextEditorDecorationType({
-    backgroundColor: new vscode.ThemeColor('list.focusHighlightForeground'),
-})
+import { parseCloudWatchLogsUri, createURIFromArgs } from '../cloudWatchLogsUtils'
 
 export class LogStreamDocumentProvider implements vscode.TextDocumentContentProvider, vscode.DefinitionProvider {
     // Expose an event to signal changes of _virtual_ documents
@@ -79,24 +75,6 @@ export class LogStreamDocumentProvider implements vscode.TextDocumentContentProv
         } catch (err) {
             throw new Error(`cwl: Error determining definition for content in ${document.fileName}`)
         }
-    }
-}
-
-export async function highlightDocument(registry: LogStreamRegistry, uri: vscode.Uri): Promise<void> {
-    const textEditor = registry.getTextEditor(uri)
-    const logData = registry.getLogData(uri)
-
-    if (!logData) {
-        throw new Error(`Missing log data in registry for uri key: ${uriToKey(uri)}. Unable to highlight`)
-    }
-
-    if (!textEditor) {
-        throw new Error(`Missing textEditor in registry for uri key: ${uriToKey(uri)}. Unable to highlight`)
-    }
-
-    if (logData.parameters.filterPattern) {
-        const ranges = findOccurencesOf(textEditor.document, logData.parameters.filterPattern)
-        textEditor.setDecorations(HIGHLIGHTER, ranges)
     }
 }
